Forking UnixLinkerTool into AndroidLinkerTool. There's some common things between them but that's mostly incidental at this point. Once we have some more platforms working we can see if there are attributes/etc we can apply everywhere (ensuring they're ignored when not used on a particular tool/etc). For now, this stuff is whack-a-mole and having them shared makes it harder to iterate on any one target.
With this our .so's for android end up with no imports and strictly the
functions we generate - an unstripped simple_mul is <500 bytes (most of
which is the source path names and junk in .comment).

Progress on #4606 - needs testing, script updates, and checks that things
like tracy still have symbols.